app/requirements.txt
flask==3.1.0
prometheus-client==0.21.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.4
pytest-cov==6.0.0
